Further Information & To Do, app description
I added the Further Information section, as well as the To Do section. I gathered the information in these sections from exploring the repo. I also updated the app description. I felt like it wasn't descriptive enough so I added why the app would be helpful and who the target audience is.